Separation the KeysConfiguration
